Scope legacy global styles (#74)
* Rewrite rem-based sizes in legacy site.

The old site's had some rules that set the root element's font size,
which would affect the computation of the rem unit. This global CSS is
now affecting the new site's styling, so this is one step towards
modularizing and scoping the old site's CSS.

* Nest legacy global styles under a class selector.

This also required wrapping many of the other pages with an extra
selector in order to ensure higher specificity.