Add meta tags for blog posts. (#311)
These are basic tags that mostly follow the guidance from
https://moz.com/blog/meta-data-templates-123

Note that this currently leaves off the <meta name="description"> and
<meta property="og:description"> tags. It seems that best practices for
SEO recommend writing a distinct description that's not automatically
generated from the page content itself. This is probably best done as a
separate field in Prismic, which we can do later once the page styling
is done.