Upgrade to Gatsby 4 and Node.js 16 (#366)
* Revert package-lock.json to lockfileVersion 1.

Lockfile version 2 is making it more difficult to upgrade packages since
it is tied to npm versions 7 and 8, but this app still uses npm 6 via
Node.js 14. Currently, upgrading packages is challenging because minor
changes to the requested dependencies cause extremely large cascading
changes to dependent packages, and it is hard to pinpoint exactly which
of the hundreds of upgraded packages is actually causing issues to
surface.

Reverting back to lockfile version 1 allows us to more easily detect
exactly what changes when a dependency is upgrade and will hopefully
facilite upgrading everything, including Node.js, to their latest
versions.

* Upgrade to gatsby 3.15.0.

* Upgrade none-sass to 6.0.1.

* Switch to node 16.20.2.

* Upgrade package-lock.json to lockfileVersion 2.

* Upgrade gatsby to 4.25.7.

* Update code for gatsby-plugin-prismic-previews v5.

https://github.com/prismicio/prismic-gatsby/blob/v6.0.0/docs/migration-guide-v4-to-v5.md

This upgrades our code to match the changes needed for v5 of
gatsby-plugin-prismic-previews. The three major changes are:

- Removing the manually imported Prismic preview stylesheet
- Centralizing the Prismic preview data into the PrismicPreviewProvider,
  rather than duplicating it on every page that uses the
  withPrismicPreview HoCs
- Replacing the `raw` field in GraphQL queries with `richText`

* Override gatsby-source-prismic's dependency on gatsby-source-filesystem to a lower version.

gatsby-source-prismic 5.3.1 contains an important fix
(https://github.com/prismicio/prismic-gatsby/pull/532) related to
`@imgix/gatsby`, but contains another commit that unnecessarily depends
on a newer-than-necessary version of gatsby-source-filesystem due to
https://github.com/prismicio/prismic-gatsby/pull/529/files#diff-bf74e96df5fcca670d975bab0f3ff6a636e701d5394a7632b17eebdd05f330e8R54.
This version of gatsby-source-prismic is supposed to still support
Gatsby v4 and gatsby-source-filesystem v4, but due to how it declared
its dependencies, npm will try to use gatsby-source-filesystem v5, which
results in multiple versions of that dependency being used in our Gatsby
v4 app.

This adds an npm override to force gatsby-source-prismic to use the same
version of gatsby-source-filesystem as gatsby itself, which will avoid a
loud warning message when building the site.

* Add workaround for Jest.