Port 404 page to new layout and remove legacy styles (#370)
* Port 404 page to new layout.

* Delete legacy code.

* Update new site CSS to compensate for removal of legacy CSS.

Although we had tried to isolate the legacy CSS styles, some were global
styles that affected the whole site. This updates the remaining CSS
styles in order to retain the look and feel even with the legacy CSS
removed.

One style change that does not match the old site is the link styling.
This actually makes some changes to match the styling shown in the
original Figma designs, but that we were enable to make due to the reset
styles of the previous site being too difficult to override. Now that
they are gone, we can actually implement the styling suggested by the
designs.

* Remove unnecessary Sass packages, including node-sass.

Now that all .scss files have been removed, we can remove most of the
Sass-related packages. One exception is the stylelint configuration,
since it still contains rules that we'd like to enforce, and it would be
more work to switch to a different ruleset.