Use new Candid (formerly GuideStar) seal. (#384)
This is also a direct link to an image on their website, so it should
automatically update when our status changes in the future, which will
avoid having to make a code change to update this.

Also, this bumps up the image size slightly, from 60px to 80px. The text
was difficult to read at 60px, even with the old GuideStar seal, so we
want to increase it a bit. We still want it to be smaller than our own
logo in the footer, so 80px is a reasonable middle ground.